+++ src/basic/hooks/useCouponForm.ts
@@ -1,13 +1,18 @@
 import { useState } from "react";
 
+export const INITIAL_COUPON_FORM = {
+  name: "",
+  code: "",
+  discountType: "amount" as "amount" | "percentage",
+  discountValue: 0,
+};
+
 export const useCouponForm = () => {
+  /** 쿠폰 등록 폼 표시 상태 */
   const [showCouponForm, setShowCouponForm] = useState(false);
-  const [couponForm, setCouponForm] = useState({
-    name: "",
-    code: "",
-    discountType: "amount" as "amount" | "percentage",
-    discountValue: 0,
-  });
+
+  /** 쿠폰 등록 폼 데이터 */
+  const [couponForm, setCouponForm] = useState(INITIAL_COUPON_FORM);
 
   /** 쿠폰 등록 폼 초기화 */
   const resetCouponForm = () => {
